Change ordering value to string #JC19-352
