+++ libs/core/dtos/params.dto.ts
@@ -1,14 +1,3 @@
-/** Sorting dto options for user. */
-export enum SortingDto {
-  Default = '',
-  EnglishTitle = 'title_eng',
-  AiredStart = 'aired__startswith',
-  Status = 'status',
-  EnglishTitleDecs = '-title_eng',
-  AiredStartDecs = '-aired__startswith',
-  StatusDecs = '-status',
-}
-
 /** Params for query. */
 export interface ParamsDto {
 
@@ -19,5 +8,5 @@ export interface ParamsDto {
   readonly limit: string;
 
   /** Sorting type for anime list. */
-  readonly ordering: SortingDto;
+  readonly ordering: string;
 }
